Dev/Epm.LGoods.ProductManagement/Epm.LGoods.ProductManagement.Infrastructure/Repositories/PriceRepository.cs
using System;
using System.Collections.Generic;
using System.Linq;
using System.Text;
using System.Threading.Tasks;
using Epm.LGoods.ProductManagement.Core.Interfaces;
using Epm.LGoods.ProductManagement.Core.Entities;
using Epm.LGoods.ProductManagement.Infrastructure.Data;
using Microsoft.EntityFrameworkCore;

namespace Epm.LGoods.ProductManagement.Infrastructure.Repositories
{
    public class PriceRepository : IPriceRepository
    {
        private readonly ApplicationDbContext _context;

        public PriceRepository(ApplicationDbContext context)
        {
            _context = context;
        }

        public async Task AddPriceAsync(Price price)
        {
            _context.Prices.Add(price);
            await _context.SaveChangesAsync();
        }
        public async Task<Price> GetPriceByIdAsync(int id)
        {

            return await _context.Prices.FirstOrDefaultAsync(p => p.PriceId == id);
        }


        public async Task UpdatePriceAsync(Price price)
        {
            _context.Prices.Update(price);
            await _context.SaveChangesAsync();
        }

    }
}
